Modify extension to make it compatible to enable/disable configuration GigyaPost was replacing LoginPost and CreatePost Magento default actions This way was impossible to call CreatePost::execute when disabled Created class to extend CreatePost. Now GigyaPost extends LoginPost
